+++ fleets/test/test/build.rs
@@ -1,0 +1,3 @@
+fn main() {
+    canic::build!("../test-configs/root-sharding.toml");
+}